Move actions in root actions/ package to a standard/ one
