+++ bot/manager.py
@@ -1,12 +1,7 @@
-from bot.action.admin import RestartAction, EvalAction, AdminActionWithErrorMessage, AdminAction, HaltAction, \
-    GroupAdminAction
-from bot.action.answer import AnswerAction
-from bot.action.config import ConfigAction
 from bot.action.core.action import ActionGroup
 from bot.action.core.command import CommandAction
-from bot.action.core.filter import MessageAction, TextMessageAction, NoPendingAction, EditedMessageAction, PendingAction, \
-    NoForwardedMessage, VoiceMessageAction
-from bot.action.enterexit import GreetAction, LeaveAction
+from bot.action.core.filter import MessageAction, TextMessageAction, NoPendingAction, EditedMessageAction, \
+    PendingAction, NoForwardedMessage, VoiceMessageAction
 from bot.action.extra.audios import SaveVoiceAction, ListVoiceAction
 from bot.action.extra.hashtags import SaveHashtagsAction, ListHashtagsAction
 from bot.action.extra.legacypole import LegacyPoleAction
@@ -14,13 +9,18 @@ from bot.action.extra.message import ShowMessageAction
 from bot.action.extra.messages import SaveMessageAction, ListMessageAction
 from bot.action.extra.pole import SavePoleAction, ListPoleAction, ManagePoleTimezonesAction
 from bot.action.extra.random import RandomChoiceAction
-from bot.action.gapdetector import GlobalGapDetectorAction
-from bot.action.chatsettings import ChatSettingsAction
-from bot.action.internationalization import InternationalizationAction
-from bot.action.perchat import PerChatAction
-from bot.action.silence import SilenceAction
-from bot.action.toggle import GetSetFeatureAction, ToggleableFeatureAction
-from bot.action.userinfo import SaveUserAction
+from bot.action.standard.admin import RestartAction, EvalAction, AdminActionWithErrorMessage, AdminAction, HaltAction, \
+    GroupAdminAction
+from bot.action.standard.answer import AnswerAction
+from bot.action.standard.chatsettings import ChatSettingsAction
+from bot.action.standard.config import ConfigAction
+from bot.action.standard.enterexit import GreetAction, LeaveAction
+from bot.action.standard.gapdetector import GlobalGapDetectorAction
+from bot.action.standard.internationalization import InternationalizationAction
+from bot.action.standard.perchat import PerChatAction
+from bot.action.standard.silence import SilenceAction
+from bot.action.standard.toggle import GetSetFeatureAction, ToggleableFeatureAction
+from bot.action.standard.userinfo import SaveUserAction
 from bot.bot import Bot
